Fix/protected endpoint

+++ backend/routers/event_route.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-
+from utils.auth_utils import get_current_user
+from models import user_model
 from dependencies import get_db
 from models import event_model
 from schemas import event_schema, product_schema
@@ -13,7 +14,14 @@ router = APIRouter()
 
 
 @router.post("/", response_model=event_schema.Event)
-def create_event(event: event_schema.EventCreate, db: Session = Depends(get_db)):
+def create_event(  
+        event: event_schema.EventCreate, 
+        db: Session = Depends(get_db),
+        current_user: user_model.User = Depends(get_current_user)
+    ):
+    
+    if current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Operation not permitted")
     
     db_event = event_model.Event(**event.dict())
     db.add(db_event)
@@ -43,11 +51,21 @@ def get_products_event(id_event: int, db: Session = Depends(get_db)):
     return event.products
 
 @router.put("/{id_event}", response_model=event_schema.Event)
-def update_event(id_event: int, event: event_schema.EventBase, db: Session = Depends(get_db)):
+def update_event(
+    id_event: int, 
+    event: event_schema.EventBase, 
+    db: Session = Depends(get_db),
+    current_user: user_model.User = Depends(get_current_user)
+    ):
+    if current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Operation not permitted: only admins can update events")
+    
     db_event = db.query(event_model.Event).filter(event_model.Event.id == id_event).first()
     
     if not db_event:
         raise HTTPException(status_code=404, detail=NOT_FOUND)
+    if db_event.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Operation not permitted: you can only update your own events")
     
     if event.name != None:
         db_event.name = event.name
@@ -64,13 +82,23 @@ def update_event(id_event: int, event: event_schema.EventBase, db: Session = Dep
     return db_event
         
 @router.delete("/{id_event}")
-def delete_event(id_event: int, db: Session = Depends(get_db)):
+def delete_event(
+    id_event: int, 
+    db: Session = Depends(get_db),
+    current_user: user_model.User = Depends(get_current_user)
+    ):
+    if current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Operation not permitted: only admins can delete events")
+    
     db_event = db.query(event_model.Event).filter(event_model.Event.id == id_event).first()
     
     if not db_event:
-        HTTPException(status_code=404, detail=NOT_FOUND)
+        raise HTTPException(status_code=404, detail=NOT_FOUND)
+    
+    if db_event.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Operation not permitted: you can only delete your own events")
     
     db.delete(db_event)
     db.commit()
-    
-    return {"message": f"Item with {id_event} deleted sucessfully"}
+
+    return {"message": f"Item with {id_event} deleted successfully"}
